Hawkular-922 update ui-services, fix notification handling, other - update ui-services version for group trigger services - fix email-switch handling in trigger detail - remove editing of name, description because now we are editing group triggers - remove verbose server/console logging - rename hawkular-glue jar to hawkular-listener jar
