+++ console/src/main/scripts/plugins/metrics/html/directives/fieldset-notification.html
@@ -6,7 +6,8 @@
     <label class="col-sm-3 control-label" for="email-heap">Email</label>
 
     <div class="col-sm-6">
-      <input type="text" id="email-heap" class="form-control" ng-model="hkAlertEmail" ng-disabled="hkAlertEmailEnabled">
+      <input type="text" id="email-heap" class="form-control" ng-model="hkAlertEmail"
+             ng-disabled="!hkAlertEmailEnabled">
     </div>
     <div class="col-sm-3">
       <hk-switch hk-model="hkAlertEmailEnabled" class="pull-right" id="email-switch" ></hk-switch>
