+++ console/src/main/scripts/plugins/metrics/html/partials/trigger-general-setup.html
@@ -22,6 +22,7 @@
                   </select>
                 </div>
               </div>
+              <!-- Name and Description are not editable at the group level but should be at the member level
               <div class="form-group">
                 <label class="col-sm-2 control-label">Name</label>
                 <div class="col-sm-10 hk-input-text">
@@ -32,6 +33,19 @@
                 <label class="col-sm-2 control-label">Description</label>
                 <div class="col-sm-10 hk-input-text">
                   <input class="form-control" ng-model="tc.adm.trigger.description" />
+                </div>
+              </div>
+              -->
+              <div class="form-group">
+                <label class="col-sm-2 control-label">Name</label>
+                <div class="col-sm-10">
+                  <span class="hk-input-text hk-bold">{{tc.adm.trigger.name}}</span>
+                </div>
+              </div>
+              <div class="form-group">
+                <label class="col-sm-2 control-label">Description</label>
+                <div class="col-sm-10">
+                  <span class="hk-input-text hk-bold">{{tc.adm.trigger.description}}</span>
                 </div>
               </div>
               <div class="form-group">
