+++ console/src/main/scripts/plugins/metrics/html/triggers/availability.html
@@ -45,7 +45,9 @@
                                      hk-title-unmet="Only when it is down for more than:">
               </hk-fieldset-dampening>
 
-              <hk-fieldset-notification hk-alert-email="tc.adm.trigger.email"></hk-fieldset-notification>
+              <hk-fieldset-notification
+                    hk-alert-email="tc.adm.trigger.email"
+                    hk-alert-email-enabled="tc.adm.trigger.emailEnabled"></hk-fieldset-notification>
             </form>
 
             <div class="form-actions hk-form-actions-separator text-right">
