+++ console/src/main/scripts/plugins/metrics/html/triggers/event.html
@@ -39,7 +39,9 @@
                   tc.adm.trigger.conditionContext.description : "Event"}}
               </fieldset>
 
-              <hk-fieldset-notification hk-alert-email="tc.adm.trigger.email"></hk-fieldset-notification>
+              <hk-fieldset-notification
+                    hk-alert-email="tc.adm.trigger.email"
+                    hk-alert-email-enabled="tc.adm.trigger.emailEnabled"></hk-fieldset-notification>
             </form>
 
             <div class="form-actions hk-form-actions-separator text-right">
