+++ console/src/main/scripts/plugins/metrics/html/triggers/range-percent.html
@@ -75,7 +75,9 @@
               </fieldset>
               <hk-fieldset-dampening hk-duration="tc.adm.trigger.evalTimeSetting"></hk-fieldset-dampening>
 
-              <hk-fieldset-notification hk-alert-email="tc.adm.trigger.email"></hk-fieldset-notification>
+              <hk-fieldset-notification
+                    hk-alert-email="tc.adm.trigger.email"
+                    hk-alert-email-enabled="tc.adm.trigger.emailEnabled"></hk-fieldset-notification>
             </form>
 
             <div class="form-actions hk-form-actions-separator text-right">
