+++ console/src/main/scripts/plugins/metrics/ts/alerts.ts
@@ -143,7 +143,6 @@ module HawkularMetrics {
       let triggerPromise = this.loadTrigger(triggerId);
 
       this.$q.all(triggerPromise).then(() => {
-        console.log('>>> ADM SET: ' + JSON.stringify(this.adm));
         this.admBak = angular.copy(this.adm);
         this.isSettingChange = false;
 
@@ -153,7 +152,6 @@ module HawkularMetrics {
       this.$scope.$watch(angular.bind(this, () => {
         return this.adm;
       }), () => {
-        console.log('>>> ADM: ' + JSON.stringify(this.adm));
         this.isSettingChange = !angular.equals(this.adm, this.admBak);
       }, true);
