+++ console/src/main/scripts/plugins/metrics/ts/triggers/thresholdTrigger.ts
@@ -34,9 +34,7 @@ module HawkularMetrics {
           this.adm.trigger = {};
 
           // updateable
-          this.adm.trigger['description'] = triggerData.trigger.description;
           this.adm.trigger['enabled'] = triggerData.trigger.enabled;
-          this.adm.trigger['name'] = triggerData.trigger.name;
           this.adm.trigger['severity'] = triggerData.trigger.severity;
 
           this.adm.trigger['conditionThreshold'] = triggerData.conditions[0].threshold;
@@ -46,7 +44,6 @@ module HawkularMetrics {
               this.adm.trigger[triggerAction.actionPlugin] = triggerAction.actionId;
             });
           }
-
           if ( this.adm.trigger['email'] === undefined || this.adm.trigger['email'] === null ) {
             this.adm.trigger['emailEnabled'] = false;
             this.adm.trigger['email'] = this.$rootScope.userDetails.email;
@@ -57,9 +54,12 @@ module HawkularMetrics {
           this.adm.trigger['evalTimeSetting'] = super.getEvalTimeSetting(triggerData.dampenings[0].evalTimeSetting);
 
           // presentation
+          // note: name, description not updateable at group level, should/will be at member level
           this.adm.trigger['context'] = triggerData.trigger.context;
           this.adm.trigger['conditionContext'] = triggerData.conditions[0].context;
           this.adm.trigger['conditionOperator'] = triggerData.conditions[0].operator;
+          this.adm.trigger['description'] = triggerData.trigger.description;
+          this.adm.trigger['name'] = triggerData.trigger.name;
         });
 
       return [triggerPromise];
@@ -69,20 +69,14 @@ module HawkularMetrics {
 
       let updatedFullTrigger = angular.copy(this.fullTrigger);
       updatedFullTrigger.trigger.enabled = this.adm.trigger.enabled;
-      updatedFullTrigger.trigger.name = this.adm.trigger.name;
-      updatedFullTrigger.trigger.description = this.adm.trigger.description;
       updatedFullTrigger.trigger.severity = this.adm.trigger.severity;
 
       // manipulate the TriggerAction Set appropriately
       if ( this.adm.trigger.emailEnabled ) {
-        console.log('>>>> email enabled ' + JSON.stringify(updatedFullTrigger.trigger.actions));
         updatedFullTrigger.trigger.actions = this.updateAction(
           updatedFullTrigger.trigger.actions, 'email', this.adm.trigger.email, null);  // TODO: properties
-        console.log('>>>> email enabled ' + JSON.stringify(updatedFullTrigger.trigger.actions));
       } else {
-        console.log('>>>> email NOT enabled ' + JSON.stringify(updatedFullTrigger.trigger.actions));
         updatedFullTrigger.trigger.actions = this.removeAction( updatedFullTrigger.trigger.actions, 'email' );
-        console.log('>>>> email NOT enabled ' + JSON.stringify(updatedFullTrigger.trigger.actions));
       }
 
       // When using AutoResolve the settings are implicit. We use the same dampening as for Firing mode.
