+++ pom.xml
@@ -89,7 +89,7 @@
     <version.gnu.getopt>1.0.13</version.gnu.getopt>
     <version.org.hawkular.accounts>2.0.21.Final</version.org.hawkular.accounts>
     <version.org.hawkular.agent>0.15.7.Final-SRC-revision-f486b356c1cdee2e53982da9b4dfb2f590b24c54</version.org.hawkular.agent>
-    <version.org.hawkular.alerts>0.9.0.Final-SNAPSHOT</version.org.hawkular.alerts>
+    <version.org.hawkular.alerts>0.9.0.Final-SRC-revision-cc518a4d133e04c95bd43f1a76d358e5d41dbfe4</version.org.hawkular.alerts>
     <version.org.hawkular.cmdgw>0.10.11.Final-SRC-revision-6858e69c9543587bcc15da618c6cec80b08dd108</version.org.hawkular.cmdgw>
     <version.org.hawkular.commons>0.3.5.Final</version.org.hawkular.commons>
     <version.org.hawkular.inventory>0.13.0.Final-SRC-revision-8030a5aa969e3ff5bab534112ea2b5b91c8088d7</version.org.hawkular.inventory>
